packages/data_producer/pyproject.toml
[project]
name = "src"
version = "0.1.0"
requires-python = ">=3.13"

[project.scripts]
prod = "run_scripts:production"
dev = "run_scripts:development"
producer = "run_scripts:run_producer"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python-dotenv = "^1.1.0"
paho-mqtt = "^2.1.0"
pymongo = "^4.13.0"
